pyproject.toml
[project]
name = "etl-workshop-3"
version = "0.1.0"
description = "Given the 5 CSV files which have information about happiness score in different"
authors = [
    {name = "countries, train a regression machine learning model to predict the happiness score."}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio (>=4.9.0,<5.0.0)",
    "asttokens (>=3.0.0,<4.0.0)",
    "build (>=1.2.2.post1,<2.0.0)",
    "cachecontrol (>=0.14.3,<0.15.0)",
    "catboost (>=1.2.8,<2.0.0)",
    "certifi (>=2025.4.26,<2026.0.0)",
    "charset-normalizer (>=3.4.2,<4.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "comm (>=0.2.2,<0.3.0)",
    "contourpy (>=1.3.2,<2.0.0)",
    "cycler (>=0.12.1,<0.13.0)",
    "debugpy (>=1.8.14,<2.0.0)",
    "decorator (>=5.2.1,<6.0.0)",
    "distlib (>=0.3.9,<0.4.0)",
    "dulwich (>=0.22.8,<0.23.0)",
    "executing (>=2.2.0,<3.0.0)",
    "fastjsonschema (>=2.21.1,<3.0.0)",
    "filelock (>=3.18.0,<4.0.0)",
    "findpython (>=0.6.3,<0.7.0)",
    "fonttools (>=4.58.0,<5.0.0)",
    "graphviz (>=0.20.3,<0.21.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "httpcore (>=1.0.9,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "idna (>=3.10,<4.0)",
    "installer (>=0.7.0,<0.8.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "ipython-pygments-lexers (>=1.1.1,<2.0.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.1.0,<5.0.0)",
    "jedi (>=0.19.2,<0.20.0)",
    "joblib (>=1.5.0,<2.0.0)",
    "jupyter-client (>=8.6.3,<9.0.0)",
    "jupyter-core (>=5.7.2,<6.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "kiwisolver (>=1.4.8,<2.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "matplotlib-inline (>=0.1.7,<0.2.0)",
    "more-itertools (>=10.7.0,<11.0.0)",
    "msgpack (>=1.1.0,<2.0.0)",
    "narwhals (>=1.40.0,<2.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "packaging (>=25.0,<26.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "parso (>=0.8.4,<0.9.0)",
    "patsy (>=1.0.1,<2.0.0)",
    "pbs-installer (>=2025.5.17,<2026.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pkginfo (>=1.12.1.2,<2.0.0.0)",
    "platformdirs (>=4.3.8,<5.0.0)",
    "plotly (>=6.1.1,<7.0.0)",
    "prompt-toolkit (>=3.0.51,<4.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pure-eval (>=0.2.3,<0.3.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "pyparsing (>=3.2.3,<4.0.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "pywin32 (>=310,<311)",
    "pywin32-ctypes (>=0.2.3,<0.3.0)",
    "pyzmq (>=26.4.0,<27.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "stack-data (>=0.6.3,<0.7.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "threadpoolctl (>=3.6.0,<4.0.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "tornado (>=6.5.1,<7.0.0)",
    "traitlets (>=5.14.3,<6.0.0)",
    "trove-classifiers (>=2025.5.9.12,<2026.0.0.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "urllib3 (>=2.4.0,<3.0.0)",
    "virtualenv (>=20.31.2,<21.0.0)",
    "wcwidth (>=0.2.13,<0.3.0)",
    "xgboost (>=3.0.1,<4.0.0)",
    "zstandard (>=0.23.0,<0.24.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pycountry (>=24.6.1,<25.0.0)",
    "pycountry-convert (>=0.7.2,<0.8.0)"
]

[tool.poetry]
packages = [{include = "ntlg72"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
